preliminary dashboard styling. refactored dashboard related components because the redux store was still buggy for joined data.
