+++ frontend/public/index.html
@@ -1,10 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>Simple React App</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
+
+<head>
+  <meta charset="utf-8" />
+  <title>Simple React App</title>
+  <script src="https://kit.fontawesome.com/e9c5de4366.js" crossorigin="anonymous"></script>
+
+</head>
+
+<body>
+  <div id="root"></div>
+</body>
+
 </html>
