+++ frontend/src/components/Main/Chatroom/Chatroom.css
@@ -1,0 +1,4 @@
+#chatroom__container {
+    width: 80%;
+    margin: 0 10px 20px 20px;
+}
