+++ frontend/src/components/Main/Chatroom/Chatroom.js
@@ -1,11 +1,13 @@
 import { useParams } from "react-router-dom";
 
+import './Chatroom.css';
+
 function Chatroom() {
     const { bookClubId, chatType } = useParams();
 
     return (
         <>
-            <div>
+            <div id="chatroom__container">
                 <h3>
                     My Chatroom
                 </h3>
