+++ frontend/src/components/Main/DetailsBar/DetailsBar.css
@@ -1,0 +1,7 @@
+#details__container {
+    width: 40%;
+    margin: 20px;
+    padding: 0 20px;
+    background-color: var(--sidebar-bg);
+    border-radius: 12px;
+}
