+++ frontend/src/components/Main/Main.css
@@ -1,0 +1,6 @@
+#main__container {
+    position: relative;
+    width: 100%;
+    display: flex;
+    margin: 5px;
+}
