+++ frontend/src/components/Main/Main.js
@@ -1,0 +1,17 @@
+import { Route, Switch } from 'react-router-dom';
+
+import Chatroom from './Chatroom/Chatroom';
+import DetailsBar from './DetailsBar/DetailsBar';
+
+
+function Main() {
+    return (
+        <>
+            <Chatroom />
+            <DetailsBar />
+
+        </>
+    )
+}
+
+export default Main;
