+++ frontend/src/components/Navigation/Navigation.css
@@ -1,0 +1,24 @@
+#nav__container {
+    display: flex;
+    padding-left: 20px;
+}
+
+#nav__title {
+    color: var(--main-color);
+    font-size: var(--logo-size);
+    font-weight: 500;
+    text-transform: lowercase;
+    letter-spacing: -0.5px;
+}
+
+#nav__links {
+    margin: 9px 50px 0;
+}
+
+#nav__links a {
+    margin: 18px;
+    color: var(--main-color);
+    font-size: var(--h5-size);
+    font-weight: 500;
+    text-decoration: none;
+}
