+++ frontend/src/components/Sidebar/Sidebar.css
@@ -1,0 +1,16 @@
+#sidebar__container {
+    position: relative;
+    box-sizing: border-box;
+    /* top: -18px; */
+    min-width: 230px;
+    height: 100%;
+    padding: 15px 18px;
+    background-color: var(--sidebar-bg);
+    border-top-left-radius: 12px;
+    border-bottom-left-radius: 12px;
+}
+
+.sidebar__title {
+    font-size: var(--h4-size);
+    font-weight: 600;
+}
