+++ frontend/src/index.css
@@ -1,5 +1,51 @@
-/* TODO Add site wide styles */
-div {
-    border: 1px solid black;
-    margin: 30px;
+@import url("https://use.typekit.net/qqr7rbe.css");
+
+:root {
+    --body-font: aktiv-grotesk, sans-serif;
+    --main-color: rgba(255, 255, 255, 1);
+    --background-color: #e17e7d;
+    --dash-bg: rgba(255, 255, 255, 0.09);
+    --sidebar-bg: rgba(255, 255, 255, 0.18);
+    --body-height: calc(100vh - 110px);
+    --body-width: calc(100vw - 75px);
+
+    --body-size: 16px;
+    --logo-size: 1.8em;
+    --h1-size: 2em;
+    --h2-size: 1.8em;
+    --h3-size: 1.6em;
+    --h4-size: 1.3em;
+    --h5-size: 1.2em;
+}
+
+body {
+    position: absolute;
+    /* height: var(--body-height);
+    width: var(--body-width); */
+    color: var(--main-color);
+    font-family: var(--body-font);
+    font-size: 15px;
+    background: no-repeat fixed url("./assets/background.jpg");
+    margin: 25px 40px;
+    display: flex;
+}
+
+h1 {
+    font-size: var(--h1-size);
+}
+
+h2 {
+    font-size: var(--h2-size);
+}
+
+h3 {
+    font-size: var(--h3-size);
+}
+
+h4 {
+    font-size: var(--h4-size);
+}
+
+h5 {
+    font-size: var(--h5-size);
 }
